+++ cqrs/CHANGELOG.md
@@ -1,6 +1,6 @@
 # master
 
-* No changes yet
+* Change signature of `Entity.id()`
 
 # [[0.1.0] 2019-02-01](https://github.com/cq-rs/cqrs/releases/tag/cqrs-0.1.0)